backend/requirements.txt
fastapi==0.135.2
psycopg2-binary==2.9.10
uvicorn==0.42.0
sqlalchemy==2.0.48
pydantic==2.12.5
python-jose[cryptography]==3.5.0
bcrypt==4.0.1
python-multipart==0.0.22
reportlab==4.4.10
pandas==3.0.1
pillow==12.1.1
numpy==2.4.3
slowapi==0.1.9
pypdf==5.1.0
anthropic==0.40.0
